+++ packages/@jsii/check-node/NOTICE
@@ -1,0 +1,2 @@
+jsii
+Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
