+++ packages/@jsii/check-node/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@jsii/node-check",
+  "name": "@jsii/check-node",
   "version": "0.0.0",
   "description": "Checks for supported node versions",
   "license": "Apache-2.0",
@@ -14,13 +14,18 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/aws/jsii.git",
-    "directory": "packages/@jsii/node-checl"
+    "directory": "packages/@jsii/check-node"
   },
   "engines": {
     "node": ">= 10.3.0"
   },
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
+  "exports": {
+    ".": "./lib/index.js",
+    "./package.json": "./package.json",
+    "./run": "./lib/run.js"
+  },
   "scripts": {
     "build": "tsc --build && npm run lint",
     "watch": "tsc --build -w",
