+++ packages/@jsii/check-node/run.js
@@ -1,0 +1,3 @@
+// For compatibility with node < 12.7.0, where "exports" from package.json is
+// not honored. This can be dropped once we drop support for node < 12.7.0.
+require('./lib/run');
