+++ packages/@jsii/check-node/src/run.ts
@@ -1,0 +1,6 @@
+import { checkNode } from './index';
+
+checkNode();
+
+// So this is a valid ESM module, too.
+export = {};
