+++ packages/@jsii/check-node/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../../tsconfig-base",
+  "compilerOptions": {
+    "outDir": "./lib",
+    "rootDir": "./src",
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": ["jest.config.ts"],
+}
