--- packages/@jsii/node-check/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../../tsconfig-base",
-  "include": ["**/*.ts"],
-  "exclude": ["jest.config.ts"],
-}
