+++ packages/@jsii/runtime/package.json
@@ -35,7 +35,7 @@
   },
   "dependencies": {
     "@jsii/kernel": "^0.0.0",
-    "@jsii/node-check": "0.0.0",
+    "@jsii/check-node": "0.0.0",
     "@jsii/spec": "^0.0.0"
   },
   "devDependencies": {
