+++ packages/jsii-config/package.json
@@ -31,7 +31,7 @@
     "typescript": "~3.9.9"
   },
   "dependencies": {
-    "@jsii/node-check": "0.0.0",
+    "@jsii/check-node": "0.0.0",
     "@jsii/spec": "^0.0.0",
     "inquirer": "^7.3.3",
     "yargs": "^16.2.0"
