+++ packages/jsii-config/tsconfig.json
@@ -5,7 +5,7 @@
   ],
   "exclude": ["jest.config.ts"],
   "references": [
-    { "path": "../@jsii/node-check" },
+    { "path": "../@jsii/check-node" },
     { "path": "../@jsii/spec" },
   ],
 }
