+++ packages/jsii-diff/package.json
@@ -33,7 +33,7 @@
     "package": "package-js"
   },
   "dependencies": {
-    "@jsii/node-check": "0.0.0",
+    "@jsii/check-node": "0.0.0",
     "@jsii/spec": "^0.0.0",
     "fs-extra": "^9.1.0",
     "jsii-reflect": "^0.0.0",
