+++ packages/jsii-pacmak/bin/jsii-pacmak.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
-import * as yargs from 'yargs';
+import '@jsii/check-node/run';
 
-import '@jsii/node-check';
+import * as yargs from 'yargs';
 
 import { pacmak, configureLogging, TargetName } from '../lib';
 import { debug } from '../lib/logging';
