+++ packages/jsii-pacmak/package.json
@@ -37,7 +37,7 @@
     "package": "package-js"
   },
   "dependencies": {
-    "@jsii/node-check": "0.0.0",
+    "@jsii/check-node": "0.0.0",
     "@jsii/spec": "^0.0.0",
     "clone": "^2.1.2",
     "codemaker": "^0.0.0",
