+++ packages/jsii-pacmak/tsconfig.json
@@ -3,7 +3,7 @@
   "include": ["**/*.ts"],
   "exclude": ["jest.config.ts", "test/generated-code/examples/*"],
   "references": [
-    { "path": "../@jsii/node-check" },
+    { "path": "../@jsii/check-node" },
     { "path": "../@jsii/spec" },
     { "path": "../codemaker" },
     { "path": "../jsii" },
