+++ packages/jsii-reflect/package.json
@@ -34,7 +34,7 @@
     "package": "package-js"
   },
   "dependencies": {
-    "@jsii/node-check": "0.0.0",
+    "@jsii/check-node": "0.0.0",
     "@jsii/spec": "^0.0.0",
     "colors": "^1.4.0",
     "fs-extra": "^9.1.0",
