+++ packages/jsii-rosetta/bin/jsii-rosetta.ts
@@ -1,8 +1,8 @@
+import '@jsii/check-node/run';
+
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import * as yargs from 'yargs';
-
-import '@jsii/node-check';
 
 import {
   TranslateResult,
