+++ packages/jsii-rosetta/tsconfig.json
@@ -9,7 +9,7 @@
     "jest.config.ts",
   ],
   "references": [
-    { "path": "../@jsii/node-check" },
+    { "path": "../@jsii/check-node" },
     { "path": "../@jsii/spec" },
     { "path": "../jsii" },
   ],
