+++ packages/jsii/bin/jsii.ts
@@ -1,9 +1,9 @@
+import '@jsii/check-node/run';
+
 import * as log4js from 'log4js';
 import * as path from 'path';
 import * as process from 'process';
 import * as yargs from 'yargs';
-
-import '@jsii/node-check';
 
 import { Compiler } from '../lib/compiler';
 import { configureCategories } from '../lib/jsii-diagnostic';
