+++ packages/jsii/tsconfig.json
@@ -3,7 +3,7 @@
   "include": ["**/*.ts"],
   "exclude": ["test/negatives/*", "jest.config.ts"],
   "references": [
-    { "path": "../@jsii/node-check" },
+    { "path": "../@jsii/check-node" },
     { "path": "../@jsii/spec" },
   ]
 }
