+++ tsconfig.json
@@ -12,6 +12,7 @@
     { "path": "packages/jsii-rosetta" },
     { "path": "packages/jsii" },
     { "path": "packages/oo-ascii-tree" },
+    { "path": "packages/@jsii/check-node" },
     { "path": "packages/@jsii/dotnet-runtime-test" },
     { "path": "packages/@jsii/dotnet-runtime" },
     { "path": "packages/@jsii/integ-test" },
